requirements.txt
fire
tensorflow
google-cloud-storage
google-api-python-client
tqdm